Extract some logic to SmogProcessor.ts, and make TypeScript server, executed with 'ts-node server.ts'.
